refactor(element): move all decorators to a "decorators" directory
